Fixed some ui-slider things that got missed

+++ src/ui/ui-slider.hpp
@@ -27,7 +27,7 @@ class UI_Slider : public UI_Base, public Event_Handler
 	const int max_value;
 	const int min_value;
 	const int difference;
-	const int pos_fraction;
+	const double pos_fraction;
 	const int pos_count;
 
 	bool mouse_down;
